Refactor and enhance error handling in script functions
- src/botlib/l_script.c: Added GCC diagnostic pragmas to suppress warnings for format-nonliteral in ScriptError and ScriptWarning functions.
- src/client/cl_jpeg.c: Updated FS_ReadFile call to use va for filename formatting.
- src/client/cl_main.c: Simplified Cvar_VariableString calls for map configuration.
- src/client/mp3/csbt.c: Updated function signatures to include void for clarity.
- src/qcommon/net_ip.c: Cleaned up includes and organized header files for better compatibility.
- src/qcommon/q_math.c and q_shared.h: Changed VectorRotate function signature for consistency.
- src/renderers: Various updates to function signatures and type safety improvements across OpenGL and Vulkan renderers.
- Added new noise functions in rendercommon for enhanced noise generation capabilities.

+++ src/renderers/openglrenderer/tr_shadows.c
@@ -92,7 +92,7 @@ static void R_CalcShadowEdges( void ) {
 			// if it doesn't share the edge with another front facing
 			// triangle, it is a sil edge
 			if ( sil_edge ) {
-				if ( tess.numIndexes > ARRAY_LEN( tess.indexes ) - 6 ) {
+				if ( (size_t)tess.numIndexes > ARRAY_LEN( tess.indexes ) - 6 ) {
 					i = tess.numVertexes;
 					break;
 				}
